Add support for Windows platform (#137) (#136)
* Template for windows port

* Make core code to be windows compliant

* Implement Windows functions and fix compilation warnings

* Add Windows build on Github actions

* Add windows examples

* Fix makefile compatibility with windows

* Fix include order in windows network system

* Fix example warnings

* Update README.md with Windows support

* Properly set environment variables

* Build static on windows

* Avoid possible loss of data in type conversion

* Add z_time_now_as_str for windows

* Update Windows examples with z_str_t

* Add missing time include

* Avoid possible loss of data in type conversion

* Random uint64 instead of time

* Run integration tests on windows

* Avoid possible loss of data in type conversion

* Trigger tests in CI using ctest

* Enable testing by default again

+++ src/protocol/codec.c
@@ -49,7 +49,7 @@ int8_t _z_period_decode_na(_z_period_t *p, _z_zbuf_t *buf) {
 int8_t _z_period_decode(_z_period_t *p, _z_zbuf_t *buf) { return _z_period_decode_na(p, buf); }
 
 /*------------------ uint8 -------------------*/
-int8_t _z_enum_encode(_z_wbuf_t *wbf, int en) { return _z_wbuf_write(wbf, en); }
+int8_t _z_enum_encode(_z_wbuf_t *wbf, int en) { return _z_wbuf_write(wbf, (uint8_t)en); }
 
 int8_t _z_enum_decode(int *en, _z_zbuf_t *zbf) {
     int8_t ret = _Z_RES_OK;
